feat: add encode direction (clinical facts -> TCR raw codes) with round-trip comparison
Adds the reverse of the existing decoder: TCREncoder and apply_ssf_encode_profile
turn clinical labels back into TCR Longform raw codes, plus compare_roundtrip /
export_roundtrip_report to decode-then-reencode a registry file and diff the
result against the original.

Architecture:
- tcr_decoder/codemap.py: CodeMap is a single bidirectional code<->label
  source of truth. _map_decode() now delegates to it, so every existing
  lookup-table decoder gets .encode() for free with zero behavior change.
- tcr_decoder/encoders.py: bespoke encoders for composite/scaled/bitmask
  fields (ER/PR, HER2, Ki-67, Nottingham, AFP, EGFR, RAS, ...) that mirror
  each decoder's exact output format, including its own passthrough
  leniency for out-of-spec codes (decode never crashes, so encode doesn't
  either).
- tcr_decoder/encoder.py: TCREncoder reconstructs SSF1-10 for all 11 cancer
  profiles plus the structural fields with a code table in this codebase
  (AJCC, surgery-type codes, EBRT bitmask, LN_POSITI). Fields decoded by
  trusting the input file's own pre-existing column, or an external
  cancer_registry_mapping.py, are explicitly reported as unencoded rather
  than guessed.
- tcr_decoder/roundtrip.py: compare_roundtrip() and export_roundtrip_report()
  for validating a registry export's round-trip fidelity end to end.

Also fixes two bugs found while building this:
- add_er_pr_percent() crashed under pandas 2.x when an all-integer-percent
  column inferred as int64 and a later 0.5/0.0 override tried to upcast it.
- breast SSF8/SSF9 (Paget's disease, LVI) had decoder=None in the profile
  registry despite _decode_paget/_decode_lvi existing, so apply_ssf_profile
  silently used generic numeric passthrough instead of the real decoders.

Flags one further pre-existing bug for follow-up (documented, not fixed
here since the blast radius touches tests/docs/other modules): lung SSF3
(ECOG/KPS) and the unrelated KPSECOG field both write to the
'Performance_Status' column in TCRDecoder.decode(), so KPSECOG silently
wins -- see the KNOWN ISSUE comment in core.py.

224 new tests (all passing); full existing suite unaffected (2 failures
are pre-existing pandas/pytest version issues, confirmed via git stash
before this branch's changes).

+++ tcr_decoder/derived.py
@@ -215,8 +215,12 @@ def add_er_pr_percent(df: pd.DataFrame) -> pd.DataFrame:
         col_str = df[col].fillna('').astype(str)
 
         # Decimal-aware extraction: captures '10', '10.5', '0.5', '100'.
+        # Forced to float64 (not just to_numeric's inferred dtype) because
+        # an all-integer-percent column infers as int64, and pandas >=2
+        # raises on assigning the float overrides below (0.5, 0.0) into
+        # an int64 column instead of silently upcasting it.
         pct = col_str.str.extract(r'([0-9]+(?:\.[0-9]+)?)\s*%', expand=False)
-        df[f'{receptor}_Percent'] = pd.to_numeric(pct, errors='coerce')
+        df[f'{receptor}_Percent'] = pd.to_numeric(pct, errors='coerce').astype('float64')
 
         # Convention overrides
         lt1 = col_str.str.contains('<1%', na=False)
